pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]

[project]
classifiers = [
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
]
name = "nomad-distribution"
description = "nomad distribution template"
dynamic = ["version"]
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
dependencies = ["nomad-lab[parsing, infrastructure]>=1.3.4", "setuptools"]


[project.urls]
repository = "https://github.com/FAIRmat-NFDI/nomad-distro-template"

[project.optional-dependencies]
plugins = [
  "perovskite-solar-cell-database @ git+https://github.com/FAIRmat-NFDI/nomad-perovskite-solar-cells-database.git@v0.1.1",
  "nomad-porous-materials @ git+https://github.com/FAIRmat-NFDI/nomad-porous-materials.git@522f4a3208077f534f1c5e886527ee2104283d0b",
  "nomad-aitoolkit @ git+https://github.com/FAIRmat-NFDI/nomad-aitoolkit.git@b203f8eb28dc4b6771a39d5c5b7ad3d9d8583ac8",
  "nomad-simulations==0.3.1",
  "pynxtools[convert]==0.9.3",
  "nomad-schema-plugin-run>=1.0.1",
  "nomad-schema-plugin-simulation-workflow>=1.0.1",
  "nomad-normalizer-plugin-bandstructure>=1.0",
  "nomad-normalizer-plugin-dos>=1.0",
  "nomad-normalizer-plugin-soap>=1.0",
  "nomad-normalizer-plugin-spectra>=1.0",
  "nomad-normalizer-plugin-system>=1.0",
  "nomad-normalizer-plugin-simulation-workflow>=1.0",
  "eelsdbconverter @ git+https://github.com/nomad-coe/nomad-parser-eelsdb.git@entry-point",
  "nomad-parser-plugins-electronic>=1.0",
  "nomad-parser-plugins-atomistic>=1.0",
  "nomad-parser-plugins-workflow>=1.0",
  "nomad-parser-plugins-database>=1.0",
]
jupyter = []
dev = ["nomad-lab[parsing, infrastructure, dev]>=1.3.4"]

[tool.uv]
extra-index-url = [
  "https://gitlab.mpcdf.mpg.de/api/v4/projects/2187/packages/pypi/simple",
]
constraint-dependencies = ["hyperspy>=1.7.6"]
prerelease = "disallow"

[tool.setuptools.packages.find]
exclude = ["gui*", "docs*", "config*"]
